Add CMake support to NymphcastPlayer
QMake is slowly being deprecated by the Qt project. From Qt 6.0 and
onwards Qt itself is already building with CMake and not with QMake
anymore and it's only a matter of time til QMake itself will be removed.

This commit introduces the CMake build system and successfully builds
the project for at least Linux